requirements.txt
beautifulsoup4==4.5.3
bs4==0.0.1
future==0.16.0
getresume==0.2.1
nltk==3.2.2
nose==1.3.7
numpy==1.13.1
python-dateutil==2.6.1
scikit-learn==0.18.2
scipy==0.19.1
six==1.10.0
sklearn==0.0
stem==1.5.4
tqdm==4.11.2
